Update main.py
Found a fatal error when passing --props pk as an argument. Results in overall_score being None, which can't be rounded. Put a work around to check if overall_score is None before attempting to run. 
Note that it runs ~10% slower than --props all, so best to just calculate everything and delete the data columns that you don't want from the output.